requirements.txt
discord
PyYAML
requests